+++ src/utils/SpectralFile.h
@@ -8,19 +8,19 @@ class Spectrum;
 class SpectrumDescriptor;
 class PR_LIB_UTILS SpectralFile {
 public:
-	SpectralFile(const std::shared_ptr<SpectrumDescriptor>& desc, uint32 width, uint32 height);
-	SpectralFile(const std::shared_ptr<SpectrumDescriptor>& desc, uint32 width, uint32 height, float* data, bool copy = false);
+	SpectralFile(const std::shared_ptr<SpectrumDescriptor>& desc, size_t width, size_t height);
+	SpectralFile(const std::shared_ptr<SpectrumDescriptor>& desc, size_t width, size_t height, float* data, bool copy = false);
 	virtual ~SpectralFile();
 
-	void set(uint32 row, uint32 column, const Spectrum& spec);
-	void get(uint32 row, uint32 column, Spectrum& spec) const;
+	void set(size_t row, size_t column, const Spectrum& spec);
+	void get(size_t row, size_t column, Spectrum& spec) const;
 
 	void save(const std::wstring& path, bool compress = true) const;
 	static SpectralFile open(const std::wstring& path, bool compressed = true);
 
 	const std::shared_ptr<SpectrumDescriptor>& descriptor() const;
-	uint32 width() const;
-	uint32 height() const;
+	size_t width() const;
+	size_t height() const;
 	float* ptr() const;
 
 private:
